Add roles  and change from rulings to moves
The first portion of the game now describes what the GM and players are
responsible, and gestures towards topics I'll need to add more on.

"Rulings" was an idea to meld OSR-syle "rulings not rules" with the move
format, which is still a thing that _could_ work but isn't really needed
for this game. Going back to a more typical AW-style move concept.